cloud-functions/getYoutubeTrailer/requirements.txt
requests==2.31.0
beautifulsoup4
selenium
functions-framework==3.5.0
flask==3.0.3
google-cloud-error-reporting==1.9.1
MarkupSafe==2.1.3
ecdsa==0.19.0
httpx==0.27.0
pydantic==2.7.4
pydantic_core==2.18.4
websockets==12.0
google-cloud-firestore==2.9.0
imdbmovies==0.1.1
